+++ src/REL/m408Dll/stage.c
@@ -1069,7 +1069,7 @@ void fn_1_10484(UnkFn10484 *arg0, HsfFace *arg1, HsfDrawData *arg2, GXColor *arg
         fn_1_10830(arg0->unk0, temp_r29->bitmap, temp_r29, (s16)temp_r27);
     }
     fn_1_10B60(arg0, temp_r31);
-    temp_r20 = (void *)(arg2->dlOfs + ((u32)lbl_1_bss_120[0]));
+    temp_r20 = (void *)(arg2->dlOfs + ((uintptr_t)lbl_1_bss_120[0]));
     GXCallDisplayList(temp_r20, arg2->dlSize);
 }
